Let simple scripts use /bin/sh, and handle whitespace paths
The scripts here are very lightweight and simple, and do not need or use
the wide range of extra functionality provided by bash over what is
found in a small and fast POSIX shell (like e.g. dash). Hence those
scripts can rely on /bin/sh instead of /bin/bash.

Furthermore, this patch adds quotes for the dirname in the scripts, in
case there are whitespaces in the paths.

Change-Id: I326e528effd93380f3019af549a9dd65c11743c3
Signed-off-by: Joakim Roubert <joakimr@axis.com>